requirements.txt
# These versions are based on May 2024
# Please update the versions to the latest ones
# You can use the following command to get the latest versions
# pip install -r requirements.txt --upgrade

elastic-transport==8.13.0
elasticsearch==8.13.1
Faker==25.2.0
pyyaml==6.0.1